tests/wpt/tests/css/css-grid/masonry/tentative/track-sizing/auto-repeat/column-auto-repeat-025.html

<!DOCTYPE html>
<html>
<title>Auto-fit repeat tracks with fixed size and auto item placement</title>
<link rel="help" href="https://drafts.csswg.org/css-grid-3">
<link rel="author" title="Alison Maher" href="mailto:almaher@microsoft.com">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
.masonry {
    display: masonry;
    item-tolerance: 0;
    grid-template-columns: repeat(auto-fit, 100px) repeat(5, 100px);
    height: 200px;
    width: 1000px;
}

.masonry > div {
    width: 100%;
    height: 100px;
}
</style>
<div class="masonry">
  <div></div>
  <div></div>
  <div></div>
  <div></div>
  <div></div>
</div>
<script>
  test(function() {
    const container = document.querySelector('.masonry');
    const computedStyle = window.getComputedStyle(container);
    assert_equals(computedStyle.getPropertyValue('grid-template-columns'),
      "100px 100px 100px 100px 100px 100px 100px 100px 100px 100px");
  });
</script>
</html>
