LayoutTests/imported/w3c/web-platform-tests/css/css-grid/masonry/tentative/track-sizing/auto-repeat/column-auto-repeat-025-expected.txt


FAIL Auto-fit repeat tracks with fixed size and auto item placement assert_equals: expected "100px 100px 100px 100px 100px 100px 100px 100px 100px 100px" but got "repeat(auto-fit, 100px) repeat(5, 100px)"
